common new_partial for tanssi and simple (not frontier)
