+++ node/Cargo.toml
@@ -22,6 +22,7 @@ serde_json = { workspace = true }
 tokio = { workspace = true }
 
 # Local
+node-common = { workspace = true }
 ccp-authorities-noting-inherent = { workspace = true, features = [ "std" ] }
 dancebox-runtime = { workspace = true, features = [ "std" ] }
 manual-xcm-rpc = { workspace = true }
